правка category_id

+++ schema.sql
@@ -37,7 +37,7 @@ CREATE TABLE lots (
   bet_step INT UNSIGNED                         NOT NULL,
   owner_id INT UNSIGNED                         NOT NULL,
   winner_id INT UNSIGNED,
-  category_id INT UNSIGNED                      NOT NULL,
+  category_id TINYINT UNSIGNED                  NOT NULL,
   FOREIGN KEY (owner_id) REFERENCES users(id),
   FOREIGN KEY (winner_id) REFERENCES users(id),
   FOREIGN KEY (category_id) REFERENCES categories(id),
